pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conductor-ai"
version = "0.1.0"
description = "AI-powered agent orchestration for automated software development"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Conductor Contributors"},
]
keywords = [
    "ai", "agents", "orchestration", "automation",
    "gemini", "developer-tools", "ci-cd", "github",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "websockets>=11.0",
    "pyyaml>=6.0",
    "aiofiles>=23.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.5",
]

[project.urls]
Homepage = "https://github.com/aryatafvizi/conductor"
Documentation = "https://github.com/aryatafvizi/conductor#readme"
Repository = "https://github.com/aryatafvizi/conductor"
Issues = "https://github.com/aryatafvizi/conductor/issues"
Changelog = "https://github.com/aryatafvizi/conductor/blob/main/CHANGELOG.md"

[project.scripts]
con = "conductor.cli:main"
conductor = "conductor.cli:main"

[tool.setuptools.packages.find]
include = ["conductor*"]

[tool.setuptools.package-data]
conductor = ["static/**/*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["conductor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
